requirements.txt
anyio==4.8.0
APScheduler==3.11.0
certifi==2024.12.14
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
python-telegram-bot==21.10
sniffio==1.3.1
tzlocal==5.2
